user id implemented to frontend and post

+++ mean-course/backend/routes/posts.js
@@ -91,17 +91,26 @@ router.put("/:id",checkAuth, multer({storage: storage}).single("image"), (req, r
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'update successfull'});
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: 'Update successfull!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   });
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    res.status(200).json({message: "Post delete!"});
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    console.log(result)
+    if (result.n > 0) {
+      res.status(200).json({message: 'Deletion successfull!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   })
 });
